requirements.txt
pandas>=1.0.0
numpy>=1.18.0
pyyaml
seaborn
matplotlib
nltk
scikit-learn